refactor Tap interaction
