+++ lib/interaction/Tap/index.js
@@ -52,67 +52,7 @@ const TapGesture = function (source, target, options) {
 
 module.exports = TapGesture
 const proto = TapGesture.prototype
+proto.isTap = true
 
-proto.bind = function () {
-  // @tapspace.interaction.Tap:bind()
-  //
-  // Bind gesture event listeners.
-  //
-  if (this.bound) {
-    return true
-  }
-  this.bound = true
-
-  const self = this
-
-  // Construct listeners
-  this.ongesturestart = (ev) => {
-    if (self.bound) {
-      self.target.element.classList.add('active-tap')
-      self.source.emit('tapstart', ev)
-    }
-  }
-  this.ongestureend = (ev) => {
-    if (self.bound) {
-      self.target.element.classList.remove('active-tap')
-      if (ev.travel <= self.options.maxTravel) {
-        self.source.emit('tapend', ev)
-        self.source.emit('tap', ev)
-      } else {
-        self.source.emit('tapcancel', ev)
-      }
-    }
-  }
-  this.ongesturecancel = (ev) => {
-    if (self.bound) {
-      self.target.element.classList.remove('active-tap')
-      self.source.emit('tapcancel', ev)
-    }
-  }
-
-  // Pass capturer options
-  const capturerOptions = { preventDefault: this.options.preventDefault }
-
-  // Bind listeners to the component capturer
-  this.capturer = this.source.capturer('gesture', capturerOptions)
-  this.capturer.on('gesturestart', this.ongesturestart)
-  this.capturer.on('gestureend', this.ongestureend)
-  this.capturer.on('gesturecancel', this.ongesturecancel)
-}
-
-proto.unbind = function () {
-  // @tapspace.interaction.Tap:unbind()
-  //
-  // Unbind capturer.
-  //
-  if (this.bound) {
-    this.bound = false
-    this.capturer.off('gesturstart', this.ongesturstart)
-    this.capturer.off('gestureend', this.ongestureend)
-    this.capturer.off('gesturecancel', this.ongesturecancel)
-    this.capturer = null
-    this.ongesturestart = null
-    this.ongestureend = null
-    this.ongesturecancel = null
-  }
-}
+proto.bind = require('./bind')
+proto.unbind = require('./unbind')
